refactor: move dates and associated structs to own package
