+++ Cargo.toml
@@ -22,3 +22,4 @@ sqlx = { version = "0.7.3", features = ["macros", "postgres", "bigdecimal"] }
 shuttle-shared-db = { version = "0.34.0", features = ["postgres"] }
 async-traits = "0.0.0"
 toml = "0.8.8"
+chrono = { version = "0.4.31", features = ["serde"] }
