+++ src/domain.rs
@@ -1,2 +1,3 @@
+pub mod dates;
 pub mod postgres_repository;
 pub mod repository;
