+++ src/domain/repository.rs
@@ -1,8 +1,7 @@
+use super::dates::Date;
 use actix_web::web;
 use anyhow::anyhow;
-use serde::Serialize;
 use shuttle_runtime::async_trait;
-use sqlx::prelude::FromRow;
 use std::sync::Mutex;
 
 pub type AppState = web::Data<Box<dyn Repository + Send + Sync>>;
@@ -37,39 +36,7 @@ pub trait Repository {
     /// * `date_id`:
     async fn get(&self, date_id: &uuid::Uuid) -> Option<Date>;
 }
-#[derive(Debug, Serialize, Clone, PartialEq, FromRow)]
-pub struct Description {
-    text: String,
-}
-#[derive(Debug, Serialize, Clone, PartialEq, FromRow)]
-/// Date storage
-///
-/// * `name`: The name of the date
-/// * `count`: The number of upvotes for the date.
-pub struct Date {
-    pub name: String,
-    pub count: i32,
-    pub id: uuid::Uuid,
-    pub description: Option<Description>,
-}
-impl Date {
-    pub fn new(name: impl Into<String>) -> Date {
-        Date {
-            name: name.into(),
-            count: 0,
-            id: uuid::Uuid::new_v4(),
-            description: None,
-        }
-    }
-    pub fn add(&mut self) {
-        self.count += 1;
-    }
-    pub fn minus(&mut self) {
-        if self.count > 0 {
-            self.count -= 1;
-        }
-    }
-}
+
 pub struct VecRepo {
     pub dates: Mutex<Vec<Date>>,
 }
